rename utility to setApplicationLocale. fix failing wdio test.
