+++ packages/terra-functional-testing/src/commands/utils/index.js
@@ -1,13 +1,13 @@
 const describeViewports = require('./describeViewports');
 const getViewports = require('./getViewports');
 const hideInputCaret = require('./hideInputCaret');
-const setLocale = require('./setLocale');
+const setApplicationLocale = require('./setApplicationLocale');
 const setViewport = require('./setViewport');
 
 module.exports = {
   describeViewports,
   getViewports,
   hideInputCaret,
-  setLocale,
+  setApplicationLocale,
   setViewport,
 };
