+++ packages/terra-functional-testing/src/commands/utils/setApplicationLocale.js
