+++ packages/terra-functional-testing/src/terra-dev-site/terra-functional-testing/wdio-testing/About.tool.mdx
@@ -162,11 +162,11 @@ it('should hide the caret', () => {
 
 ## Locale Update (Case by Case)
 
-### Terra.setLocale
-Use `Terra.setLocale` to update a terra application's locale on a case by case basis; particularly useful for testing locale changes on a deployed application. This method accepts a string containing the `locale` to update to.
+### Terra.setApplicationLocale
+Use `Terra.setApplicationLocale` to update a terra application's locale on a case by case basis; particularly useful for testing locale changes on a deployed application. This method accepts a string containing the `locale` to update to.
 
 ```js
 it('should update an application's locale, () => {
-  Terra.setLocale('en');
+  Terra.setApplicationLocale('en');
 });
 ```
